Renames poet -> log
- Renames the poet module to log
- Renames NewLogger to just New, so that you have less redundancy. i.e. log.NewLogger becomes log.New
- Renames LogLevel to just Level, so that you have less redundancy. i.e. log.LogLevel becomes log.Level

Resolves #81
