+++ detect.go
@@ -26,7 +26,7 @@ import (
 	"github.com/BurntSushi/toml"
 
 	"github.com/buildpacks/libcnb/internal"
-	"github.com/buildpacks/libcnb/poet"
+	"github.com/buildpacks/libcnb/log"
 )
 
 // DetectContext contains the inputs to detection.
@@ -88,7 +88,7 @@ func Detect(detector Detector, options ...Option) {
 		ok   bool
 	)
 	ctx := DetectContext{}
-	logger := poet.NewLogger(os.Stdout)
+	logger := log.New(os.Stdout)
 
 	ctx.Application.Path, err = os.Getwd()
 	if err != nil {
